pyproject.toml
[project]
name = "omb"
version = "0.1.0"
description = "Open Memory Benchmark"
requires-python = ">=3.11"
dependencies = [
    "datasets>=2.0",
    "typer>=0.12",
    "rich>=13",
    "rank-bm25>=0.2",
    "google-genai>=1.0.0",
    "mem0ai>=1.0.5",
    "sentence-transformers>=3.0",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "qdrant-client>=1.13",
    "fastapi[standard]>=0.1.0",
    "uvicorn>=0.30.0",
    "tiktoken>=0.1.0",
    "groq>=0.1.0",
]

[project.scripts]
omb = "memory_bench.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/memory_bench"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
